CMake: -adding DIMACS benchmark DIMACS: -adding example code to read benchmark on DIMACS file HPF: -attempted to get cap and rcap working for same arc, but no luck IBFS: -noticed major bug when run on data/BL06-gargoyle-med.max example,  computes wrong maxflow Util: -timer added

+++ src/algorithms/ibfs/ibfs.cpp
@@ -929,17 +929,17 @@ template<bool sTree> void IBFSGraph::augmentIncrements()
 }
 
 
-int IBFSGraph::computeMaxFlow()
+unsigned long long IBFSGraph::computeMaxFlow()
 {
 	return computeMaxFlow(true, false);
 }
 
-int IBFSGraph::computeMaxFlow(bool allowIncrements)
+unsigned long long IBFSGraph::computeMaxFlow(bool allowIncrements)
 {
 	return computeMaxFlow(true, allowIncrements);
 }
 
-int IBFSGraph::computeMaxFlow(bool initialDirS, bool allowIncrements)
+unsigned long long IBFSGraph::computeMaxFlow(bool initialDirS, bool allowIncrements)
 {
 	// incremental?
 	if (incIteration >= 1 && incList != NULL) {
